completely refactored code to accomodate new hands json file format.
The code is still compatible with the old format, but it also works with the new JSON format. The code implements a basic wrist filtering algorithm by choosing a wrist with "reasonable" values i.e. values
that are humanly possible. Still much work to do, but code is mostly functional.